Extract .info file generation logic
The function generating an .info file is currently not used (the code is
duplicated in Upgrade.hs, but replacing the Upgrade.hs code needs a bit
more refactoring since Upgrade.hs works with pure text, but PackageInfo
has [URL] and [Digest] in it. Additional converting between Text and
URL/Digest is actually a good thing since it provides some validation.

Next steps:
- `generate` should use the text builder
- replacing version in the download URL should be extracted into a new
  function and the replacement should probably be done on the URL data
  structure directly